Novo design. Interfaçe do cliente.
