+++ Clinica_Odonto/nbproject/project.properties
@@ -30,7 +30,8 @@ excludes=
 includes=**
 jar.compress=false
 javac.classpath=\
-    ${libs.MySQLDriver.classpath}
+    ${libs.MySQLDriver.classpath}:\
+    ${libs.absolutelayout.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -54,7 +55,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=Design.ClinicaFrame
+main.class=Design.ClinicaF
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
