+++ Clinica_Odonto/src/Design/AdmPanel.java
@@ -1,0 +1,614 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Design;
+
+import Persistence.Connect;
+import Users.*;
+import Utiliarios.DesignTools;
+import java.sql.Connection;
+import java.util.*;
+import javax.swing.DefaultComboBoxModel;
+import javax.swing.table.DefaultTableModel;
+
+/**
+ *
+ * @author LucasPc
+ */
+public class AdmPanel extends javax.swing.JPanel {
+
+    /**
+     * Creates new form AdmPanel
+     */
+    public AdmPanel() {
+        initComponents();
+    }
+    Clinica _clinica = new Clinica();
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jConsultaDialog = new javax.swing.JDialog();
+        jLabel1 = new javax.swing.JLabel();
+        txtCpfCliente = new javax.swing.JTextField();
+        jLabel2 = new javax.swing.JLabel();
+        txtRegMedic = new javax.swing.JTextField();
+        jBtMarcarConsulta = new javax.swing.JButton();
+        cbSetRegMed = new javax.swing.JComboBox();
+        jLabel = new javax.swing.JLabel();
+        jClienteDialog = new javax.swing.JDialog();
+        jLabel3 = new javax.swing.JLabel();
+        txtCpfCliente1 = new javax.swing.JTextField();
+        jLabel4 = new javax.swing.JLabel();
+        txtNomeCliente = new javax.swing.JTextField();
+        jBtAddCliente = new javax.swing.JButton();
+        jLabel5 = new javax.swing.JLabel();
+        txtTelCliente = new javax.swing.JTextField();
+        jDentistaDialog = new javax.swing.JDialog();
+        jLabel8 = new javax.swing.JLabel();
+        txtCpfMedico = new javax.swing.JTextField();
+        jLabel9 = new javax.swing.JLabel();
+        txtNomeMedico = new javax.swing.JTextField();
+        jLabel10 = new javax.swing.JLabel();
+        txtRegistroMedico = new javax.swing.JTextField();
+        jLabel11 = new javax.swing.JLabel();
+        JCBEspecialidade = new javax.swing.JComboBox();
+        jBAddDentista = new javax.swing.JButton();
+        jTabbedPane1 = new javax.swing.JTabbedPane();
+        ConsultasPanel = new javax.swing.JPanel();
+        jSeparator1 = new javax.swing.JSeparator();
+        jBtAddConsulta = new javax.swing.JButton();
+        jTextField1 = new javax.swing.JTextField();
+        jButton4 = new javax.swing.JButton();
+        jSeparator5 = new javax.swing.JSeparator();
+        jScrollPane4 = new javax.swing.JScrollPane();
+        tbConsulta = new javax.swing.JTable();
+        ClientesPanel = new javax.swing.JPanel();
+        jSeparator2 = new javax.swing.JSeparator();
+        jBtAddNCliente = new javax.swing.JButton();
+        jScrollPane2 = new javax.swing.JScrollPane();
+        tbCliente = new javax.swing.JTable();
+        GerenciaPanel = new javax.swing.JPanel();
+        jSeparator4 = new javax.swing.JSeparator();
+        jbAddDentistadg = new javax.swing.JButton();
+        jButton5 = new javax.swing.JButton();
+        jSeparator6 = new javax.swing.JSeparator();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        tbFuncionario = new javax.swing.JTable();
+        jScrollPane3 = new javax.swing.JScrollPane();
+        tbDentista = new javax.swing.JTable();
+
+        jConsultaDialog.setTitle("Nova Consulta");
+        jConsultaDialog.setMinimumSize(new java.awt.Dimension(334, 194));
+        jConsultaDialog.setResizable(false);
+
+        jLabel1.setText("CPF");
+
+        jLabel2.setText("Data");
+
+        jBtMarcarConsulta.setText("Adicionar");
+        jBtMarcarConsulta.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtMarcarConsultaActionPerformed(evt);
+            }
+        });
+
+        cbSetRegMed.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
+
+        jLabel.setText("Médico");
+
+        javax.swing.GroupLayout jConsultaDialogLayout = new javax.swing.GroupLayout(jConsultaDialog.getContentPane());
+        jConsultaDialog.getContentPane().setLayout(jConsultaDialogLayout);
+        jConsultaDialogLayout.setHorizontalGroup(
+            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jBtMarcarConsulta)
+                    .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 27, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 31, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel))
+                        .addGap(55, 55, 55)
+                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(cbSetRegMed, javax.swing.GroupLayout.PREFERRED_SIZE, 182, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(txtRegMedic, javax.swing.GroupLayout.DEFAULT_SIZE, 182, Short.MAX_VALUE)
+                                .addComponent(txtCpfCliente)))))
+                .addContainerGap(42, Short.MAX_VALUE))
+        );
+        jConsultaDialogLayout.setVerticalGroup(
+            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel1)
+                    .addComponent(txtCpfCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(txtRegMedic, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel2))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jLabel)
+                    .addComponent(cbSetRegMed, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(18, 18, 18)
+                .addComponent(jBtMarcarConsulta)
+                .addContainerGap(32, Short.MAX_VALUE))
+        );
+
+        jConsultaDialog.getAccessibleContext().setAccessibleName("Cinsu");
+
+        jClienteDialog.setTitle("Novo Cliente");
+        jClienteDialog.setMinimumSize(new java.awt.Dimension(334, 189));
+        jClienteDialog.setResizable(false);
+
+        jLabel3.setText("CPF");
+
+        jLabel4.setText("Nome");
+
+        jBtAddCliente.setText("Adicionar");
+        jBtAddCliente.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtAddClienteActionPerformed(evt);
+            }
+        });
+
+        jLabel5.setText("Telefone");
+
+        javax.swing.GroupLayout jClienteDialogLayout = new javax.swing.GroupLayout(jClienteDialog.getContentPane());
+        jClienteDialog.getContentPane().setLayout(jClienteDialogLayout);
+        jClienteDialogLayout.setHorizontalGroup(
+            jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jClienteDialogLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jClienteDialogLayout.createSequentialGroup()
+                        .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 27, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel4)
+                            .addComponent(jLabel5))
+                        .addGap(50, 50, 50)
+                        .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(txtTelCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 152, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(txtNomeCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 152, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(txtCpfCliente1, javax.swing.GroupLayout.PREFERRED_SIZE, 152, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                    .addComponent(jBtAddCliente))
+                .addContainerGap(80, Short.MAX_VALUE))
+        );
+        jClienteDialogLayout.setVerticalGroup(
+            jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jClienteDialogLayout.createSequentialGroup()
+                .addGap(14, 14, 14)
+                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(txtCpfCliente1)
+                    .addComponent(jLabel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel4)
+                    .addComponent(txtNomeCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel5)
+                    .addComponent(txtTelCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jBtAddCliente)
+                .addContainerGap(24, Short.MAX_VALUE))
+        );
+
+        jDentistaDialog.setTitle("Novo Dentista");
+        jDentistaDialog.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        jDentistaDialog.setMinimumSize(new java.awt.Dimension(334, 208));
+        jDentistaDialog.setResizable(false);
+
+        jLabel8.setText("CPF");
+
+        jLabel9.setText("Nome");
+
+        jLabel10.setText("Registro");
+
+        jLabel11.setText("Especialidade");
+
+        jBAddDentista.setText("Adicionar");
+        jBAddDentista.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBAddDentistaActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jDentistaDialogLayout = new javax.swing.GroupLayout(jDentistaDialog.getContentPane());
+        jDentistaDialog.getContentPane().setLayout(jDentistaDialogLayout);
+        jDentistaDialogLayout.setHorizontalGroup(
+            jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jDentistaDialogLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jDentistaDialogLayout.createSequentialGroup()
+                        .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel11)
+                            .addComponent(jLabel10)
+                            .addComponent(jLabel9)
+                            .addComponent(jLabel8))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addComponent(JCBEspecialidade, 0, 160, Short.MAX_VALUE)
+                            .addComponent(txtNomeMedico)
+                            .addComponent(txtCpfMedico)
+                            .addComponent(txtRegistroMedico)))
+                    .addComponent(jBAddDentista))
+                .addContainerGap(38, Short.MAX_VALUE))
+        );
+        jDentistaDialogLayout.setVerticalGroup(
+            jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jDentistaDialogLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel8)
+                    .addComponent(txtCpfMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel9)
+                    .addComponent(txtNomeMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel10)
+                    .addComponent(txtRegistroMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel11)
+                    .addComponent(JCBEspecialidade, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jBAddDentista)
+                .addContainerGap(20, Short.MAX_VALUE))
+        );
+
+        setBorder(javax.swing.BorderFactory.createBevelBorder(javax.swing.border.BevelBorder.LOWERED));
+        setMinimumSize(new java.awt.Dimension(462, 564));
+        setPreferredSize(new java.awt.Dimension(462, 564));
+
+        jTabbedPane1.setPreferredSize(new java.awt.Dimension(442, 542));
+
+        ConsultasPanel.setBackground(new java.awt.Color(255, 255, 255));
+
+        jBtAddConsulta.setText("Adicionar Consulta");
+        jBtAddConsulta.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtAddConsultaActionPerformed(evt);
+            }
+        });
+
+        jButton4.setText("Pesquisar");
+
+        jSeparator5.setOrientation(javax.swing.SwingConstants.VERTICAL);
+
+        tbConsulta.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        jScrollPane4.setViewportView(tbConsulta);
+
+        javax.swing.GroupLayout ConsultasPanelLayout = new javax.swing.GroupLayout(ConsultasPanel);
+        ConsultasPanel.setLayout(ConsultasPanelLayout);
+        ConsultasPanelLayout.setHorizontalGroup(
+            ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ConsultasPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
+                    .addGroup(ConsultasPanelLayout.createSequentialGroup()
+                        .addComponent(jBtAddConsulta, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jSeparator5, javax.swing.GroupLayout.PREFERRED_SIZE, 2, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 188, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jButton4)
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addComponent(jSeparator1))
+                .addContainerGap())
+        );
+        ConsultasPanelLayout.setVerticalGroup(
+            ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ConsultasPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton4))
+                    .addComponent(jBtAddConsulta, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jSeparator5))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, 384, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(74, Short.MAX_VALUE))
+        );
+
+        jTabbedPane1.addTab("Consultas", ConsultasPanel);
+
+        ClientesPanel.setBackground(new java.awt.Color(255, 255, 255));
+
+        jBtAddNCliente.setText("Adicionar Cliente");
+        jBtAddNCliente.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtAddNClienteActionPerformed(evt);
+            }
+        });
+
+        tbCliente.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        jScrollPane2.setViewportView(tbCliente);
+
+        javax.swing.GroupLayout ClientesPanelLayout = new javax.swing.GroupLayout(ClientesPanel);
+        ClientesPanel.setLayout(ClientesPanelLayout);
+        ClientesPanelLayout.setHorizontalGroup(
+            ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ClientesPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jSeparator2)
+                    .addGroup(ClientesPanelLayout.createSequentialGroup()
+                        .addComponent(jBtAddNCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 131, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addComponent(jScrollPane2, javax.swing.GroupLayout.DEFAULT_SIZE, 423, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        ClientesPanelLayout.setVerticalGroup(
+            ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ClientesPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jBtAddNCliente)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator2, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 384, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(74, Short.MAX_VALUE))
+        );
+
+        jTabbedPane1.addTab("Clientes", ClientesPanel);
+
+        GerenciaPanel.setBackground(new java.awt.Color(255, 255, 255));
+
+        jbAddDentistadg.setText("Adicionar Dentista");
+        jbAddDentistadg.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jbAddDentistadgActionPerformed(evt);
+            }
+        });
+
+        jButton5.setText("Adicionar Funcionario");
+
+        tbFuncionario.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        jScrollPane1.setViewportView(tbFuncionario);
+
+        tbDentista.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        jScrollPane3.setViewportView(tbDentista);
+
+        javax.swing.GroupLayout GerenciaPanelLayout = new javax.swing.GroupLayout(GerenciaPanel);
+        GerenciaPanel.setLayout(GerenciaPanelLayout);
+        GerenciaPanelLayout.setHorizontalGroup(
+            GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, GerenciaPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(jSeparator6)
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 423, Short.MAX_VALUE)
+                    .addComponent(jSeparator4, javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane3, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, GerenciaPanelLayout.createSequentialGroup()
+                        .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jbAddDentistadg, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jButton5))
+                        .addGap(0, 0, Short.MAX_VALUE)))
+                .addContainerGap())
+        );
+        GerenciaPanelLayout.setVerticalGroup(
+            GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(GerenciaPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jbAddDentistadg)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator4, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane3, javax.swing.GroupLayout.PREFERRED_SIZE, 172, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jButton5)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator6, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 181, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(54, Short.MAX_VALUE))
+        );
+
+        jTabbedPane1.addTab("Gerenciar", GerenciaPanel);
+
+        add(jTabbedPane1);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void jBtMarcarConsultaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtMarcarConsultaActionPerformed
+        Date d = new Date();
+        int reg = cbSetRegMed.getSelectedIndex();
+        if (_clinica.MarcarConsulta(new Consulta(txtCpfCliente.getText(), _clinica.Medicos.get(reg).getRegistro_M(), d))) {
+            this.jConsultaDialog.setVisible(false);
+        }
+        attTableConsulta();
+    }//GEN-LAST:event_jBtMarcarConsultaActionPerformed
+
+    private void jBtAddClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddClienteActionPerformed
+        boolean AddCliente;
+        //AddCliente = _clinica.AddCliente(new Cliente(txtCpfCliente1.getText(), txtNomeCliente.getText(), txtTelCliente.getText()));
+        //jLabel7.setText("" + AddCliente);
+        if (_clinica.AddCliente(new Cliente(txtCpfCliente1.getText(), txtNomeCliente.getText(), txtTelCliente.getText()))) {
+            this.jClienteDialog.setVisible(false);
+        } else {
+            //this.jLabel7.setText(" " + AddCliente);
+        };
+        attTableCliente();
+    }//GEN-LAST:event_jBtAddClienteActionPerformed
+
+    private void jBAddDentistaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBAddDentistaActionPerformed
+        int esp = JCBEspecialidade.getSelectedIndex();
+        esp++;
+        if (_clinica.AddMedico(new Medico(txtCpfMedico.getText(), txtNomeMedico.getText(),
+                txtRegistroMedico.getText(), esp))) {
+            this.jDentistaDialog.setVisible(false);
+        }
+        attTableDentista();
+    }//GEN-LAST:event_jBAddDentistaActionPerformed
+    public void attTableCliente() {
+        //Atualiza a tabela Cliente;
+        DefaultTableModel model = new DefaultTableModel();
+        model.setColumnIdentifiers(Cliente.colunasCliente);
+        _clinica.Clientes = _clinica.MostrarClientes();
+        for (Cliente c : _clinica.Clientes) {
+            model.addRow(new Object[]{c.getCpf(), c.getNome(), c.getTelefone()});
+        }
+        this.tbCliente.setModel(model);
+        this.tbCliente.setFillsViewportHeight(true);
+    }
+
+    private void attTableDentista() {
+        DefaultTableModel model = new DefaultTableModel();
+        model.setColumnIdentifiers(Medico.colunasMedico);
+        _clinica.Medicos = _clinica.MostrarMedicos();
+        for (Medico m : _clinica.Medicos) {
+            model.addRow(new Object[]{m.getCpf_M(), m.getNome_M(), m.getRegistro_M(), m.getStrEspecialidade()});
+        }
+        this.tbDentista.setModel(model);
+        this.tbDentista.setFillsViewportHeight(true);
+
+    }
+
+    private void attTableConsulta() {
+        DefaultTableModel model = new DefaultTableModel();
+        model.setColumnIdentifiers(Consulta.colunasConsulta);
+        _clinica.Consultas = _clinica.MostrarConsultas();
+        for (Consulta cs : _clinica.Consultas) {
+            model.addRow(new Object[]{cs.getIdConsulta(), cs.getCpfCliente(),
+                " " + cs.dadosMedico.getRegistro_M() + " (" + cs.dadosMedico.getNome_M() + ")",
+                cs.getData()});
+        }
+        this.tbConsulta.setModel(model);
+        this.tbConsulta.setFillsViewportHeight(true);
+    }
+
+    void att() {
+        attTableCliente();
+        attTableConsulta();
+        attTableDentista();
+        this.JCBEspecialidade.setModel(DesignTools.getEspModel());
+        this.cbSetRegMed.setModel(DesignTools.getMedEspModel());
+
+        //setComboBoxItems();
+    }
+    private void jBtAddConsultaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddConsultaActionPerformed
+        this.jConsultaDialog.setVisible(true);
+    }//GEN-LAST:event_jBtAddConsultaActionPerformed
+
+    private void jBtAddNClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddNClienteActionPerformed
+        this.jClienteDialog.setVisible(true);
+    }//GEN-LAST:event_jBtAddNClienteActionPerformed
+
+    private void jbAddDentistadgActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbAddDentistadgActionPerformed
+        this.jDentistaDialog.setVisible(true);
+    }//GEN-LAST:event_jbAddDentistadgActionPerformed
+    /*
+     */
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel ClientesPanel;
+    private javax.swing.JPanel ConsultasPanel;
+    private javax.swing.JPanel GerenciaPanel;
+    private javax.swing.JComboBox JCBEspecialidade;
+    private javax.swing.JComboBox cbSetRegMed;
+    private javax.swing.JButton jBAddDentista;
+    private javax.swing.JButton jBtAddCliente;
+    private javax.swing.JButton jBtAddConsulta;
+    private javax.swing.JButton jBtAddNCliente;
+    private javax.swing.JButton jBtMarcarConsulta;
+    private javax.swing.JButton jButton4;
+    private javax.swing.JButton jButton5;
+    private javax.swing.JDialog jClienteDialog;
+    private javax.swing.JDialog jConsultaDialog;
+    private javax.swing.JDialog jDentistaDialog;
+    private javax.swing.JLabel jLabel;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel10;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JLabel jLabel9;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JScrollPane jScrollPane2;
+    private javax.swing.JScrollPane jScrollPane3;
+    private javax.swing.JScrollPane jScrollPane4;
+    private javax.swing.JSeparator jSeparator1;
+    private javax.swing.JSeparator jSeparator2;
+    private javax.swing.JSeparator jSeparator4;
+    private javax.swing.JSeparator jSeparator5;
+    private javax.swing.JSeparator jSeparator6;
+    private javax.swing.JTabbedPane jTabbedPane1;
+    private javax.swing.JTextField jTextField1;
+    private javax.swing.JButton jbAddDentistadg;
+    private javax.swing.JTable tbCliente;
+    private javax.swing.JTable tbConsulta;
+    private javax.swing.JTable tbDentista;
+    private javax.swing.JTable tbFuncionario;
+    private javax.swing.JTextField txtCpfCliente;
+    private javax.swing.JTextField txtCpfCliente1;
+    private javax.swing.JTextField txtCpfMedico;
+    private javax.swing.JTextField txtNomeCliente;
+    private javax.swing.JTextField txtNomeMedico;
+    private javax.swing.JTextField txtRegMedic;
+    private javax.swing.JTextField txtRegistroMedico;
+    private javax.swing.JTextField txtTelCliente;
+    // End of variables declaration//GEN-END:variables
+
+}
