+++ Clinica_Odonto/src/Design/ClientePanel.java
@@ -1,0 +1,210 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Design;
+
+import Users.Cliente;
+import Users.Clinica;
+import Users.Consulta;
+import Utiliarios.DesignTools;
+import java.util.Date;
+import javax.swing.table.DefaultTableModel;
+
+/**
+ *
+ * @author LucasPc
+ */
+public class ClientePanel extends javax.swing.JPanel {
+
+    Cliente _cliente;
+
+    /**
+     * Creates new form ClientePanel
+     */
+    public ClientePanel() {
+        initComponents();
+        _cliente = new Cliente();
+        this.jlCpfCliente.setText(_cliente.getCpf());
+        this.cbSetRegMed.setModel(DesignTools.getMedEspModel());
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jConsultaDialog = new javax.swing.JDialog();
+        jLabel2 = new javax.swing.JLabel();
+        txtRegMedic = new javax.swing.JTextField();
+        jBtMarcarConsulta = new javax.swing.JButton();
+        cbSetRegMed = new javax.swing.JComboBox();
+        jLabel = new javax.swing.JLabel();
+        jlCpfCliente = new javax.swing.JLabel();
+        jTabbedPane2 = new javax.swing.JTabbedPane();
+        ConsultasPanel1 = new javax.swing.JPanel();
+        jSeparator3 = new javax.swing.JSeparator();
+        jBtAddConsulta1 = new javax.swing.JButton();
+        jScrollPane5 = new javax.swing.JScrollPane();
+        tbConsulta1 = new javax.swing.JTable();
+
+        jConsultaDialog.setTitle("Nova Consulta");
+        jConsultaDialog.setMinimumSize(new java.awt.Dimension(334, 194));
+        jConsultaDialog.setResizable(false);
+
+        jLabel2.setText("Data");
+
+        jBtMarcarConsulta.setText("Adicionar");
+        jBtMarcarConsulta.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtMarcarConsultaActionPerformed(evt);
+            }
+        });
+
+        cbSetRegMed.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
+
+        jLabel.setText("Médico");
+
+        javax.swing.GroupLayout jConsultaDialogLayout = new javax.swing.GroupLayout(jConsultaDialog.getContentPane());
+        jConsultaDialog.getContentPane().setLayout(jConsultaDialogLayout);
+        jConsultaDialogLayout.setHorizontalGroup(
+            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 31, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel))
+                        .addGap(55, 55, 55)
+                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addComponent(cbSetRegMed, 0, 182, Short.MAX_VALUE)
+                            .addComponent(txtRegMedic, javax.swing.GroupLayout.DEFAULT_SIZE, 182, Short.MAX_VALUE)
+                            .addComponent(jlCpfCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                    .addComponent(jBtMarcarConsulta))
+                .addContainerGap(35, Short.MAX_VALUE))
+        );
+        jConsultaDialogLayout.setVerticalGroup(
+            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jConsultaDialogLayout.createSequentialGroup()
+                .addGap(23, 23, 23)
+                .addComponent(jlCpfCliente)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(txtRegMedic, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel2))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jLabel)
+                    .addComponent(cbSetRegMed, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(18, 18, 18)
+                .addComponent(jBtMarcarConsulta)
+                .addContainerGap(49, Short.MAX_VALUE))
+        );
+
+        setBorder(javax.swing.BorderFactory.createBevelBorder(javax.swing.border.BevelBorder.LOWERED));
+        setMinimumSize(new java.awt.Dimension(462, 564));
+        setPreferredSize(new java.awt.Dimension(462, 564));
+
+        jTabbedPane2.setBackground(java.awt.SystemColor.window);
+        jTabbedPane2.setPreferredSize(new java.awt.Dimension(442, 542));
+
+        ConsultasPanel1.setBackground(new java.awt.Color(255, 255, 255));
+
+        jBtAddConsulta1.setText("Adicionar Consulta");
+        jBtAddConsulta1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtAddConsulta1ActionPerformed(evt);
+            }
+        });
+
+        tbConsulta1.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null},
+                {null, null, null, null}
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        jScrollPane5.setViewportView(tbConsulta1);
+
+        javax.swing.GroupLayout ConsultasPanel1Layout = new javax.swing.GroupLayout(ConsultasPanel1);
+        ConsultasPanel1.setLayout(ConsultasPanel1Layout);
+        ConsultasPanel1Layout.setHorizontalGroup(
+            ConsultasPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ConsultasPanel1Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(ConsultasPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane5, javax.swing.GroupLayout.DEFAULT_SIZE, 417, Short.MAX_VALUE)
+                    .addGroup(ConsultasPanel1Layout.createSequentialGroup()
+                        .addComponent(jBtAddConsulta1, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addComponent(jSeparator3))
+                .addContainerGap())
+        );
+        ConsultasPanel1Layout.setVerticalGroup(
+            ConsultasPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(ConsultasPanel1Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jBtAddConsulta1)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator3, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane5, javax.swing.GroupLayout.PREFERRED_SIZE, 388, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(70, Short.MAX_VALUE))
+        );
+
+        jTabbedPane2.addTab("Consultas", ConsultasPanel1);
+
+        add(jTabbedPane2);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void jBtAddConsulta1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddConsulta1ActionPerformed
+        this.jConsultaDialog.setVisible(true);
+    }//GEN-LAST:event_jBtAddConsulta1ActionPerformed
+
+    private void jBtMarcarConsultaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtMarcarConsultaActionPerformed
+        Date d = new Date();
+        int reg = cbSetRegMed.getSelectedIndex();
+        if (new Clinica().MarcarConsulta(new Consulta(_cliente.getCpf(), new Clinica().MostrarMedicos().get(reg).getRegistro_M(), d))) {
+            this.jConsultaDialog.setVisible(false);
+        }
+        attTbConsulta1();
+    }//GEN-LAST:event_jBtMarcarConsultaActionPerformed
+
+    void attTbConsulta1() {
+        DefaultTableModel model = new DefaultTableModel();
+        model.setColumnIdentifiers(Cliente.colConsultasCliente);
+        _cliente.consultas = _cliente.getConsultas();
+       
+        for (Consulta consulta : _cliente.consultas) {
+            model.addRow(new Object[]{consulta.getIdConsulta(), consulta.getData(), consulta.dadosMedico.getNome_M()});
+        }
+        this.tbConsulta1.setModel(model);
+        this.tbConsulta1.setFillsViewportHeight(true);
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel ConsultasPanel1;
+    private javax.swing.JComboBox cbSetRegMed;
+    private javax.swing.JButton jBtAddConsulta1;
+    private javax.swing.JButton jBtMarcarConsulta;
+    private javax.swing.JDialog jConsultaDialog;
+    private javax.swing.JLabel jLabel;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JScrollPane jScrollPane5;
+    private javax.swing.JSeparator jSeparator3;
+    private javax.swing.JTabbedPane jTabbedPane2;
+    private javax.swing.JLabel jlCpfCliente;
+    private javax.swing.JTable tbConsulta1;
+    private javax.swing.JTextField txtRegMedic;
+    // End of variables declaration//GEN-END:variables
+}
