--- Clinica_Odonto/src/Design/ClinicaFrame.java
@@ -1,873 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Design;
-
-import Persistence.Connect;
-import Users.*;
-import java.sql.Connection;
-import java.util.Date;
-import javax.swing.*;
-import javax.swing.table.DefaultTableModel;
-
-/**
- *
- * @author LucasPc
- */
-public class ClinicaFrame extends javax.swing.JFrame {
-
-    Clinica _clinica = new Clinica();
-
-    /**
-     * Creates new form ClinicaFrame
-     */
-    public ClinicaFrame() {
-
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jConsultaDialog = new javax.swing.JDialog();
-        jLabel1 = new javax.swing.JLabel();
-        txtCpfCliente = new javax.swing.JTextField();
-        jLabel2 = new javax.swing.JLabel();
-        txtRegMedic = new javax.swing.JTextField();
-        jBtMarcarConsulta = new javax.swing.JButton();
-        cbSetRegMed = new javax.swing.JComboBox();
-        jLabel = new javax.swing.JLabel();
-        jClienteDialog = new javax.swing.JDialog();
-        jLabel3 = new javax.swing.JLabel();
-        txtCpfCliente1 = new javax.swing.JTextField();
-        jLabel4 = new javax.swing.JLabel();
-        txtNomeCliente = new javax.swing.JTextField();
-        jBtAddCliente = new javax.swing.JButton();
-        jLabel5 = new javax.swing.JLabel();
-        txtTelCliente = new javax.swing.JTextField();
-        jLabel7 = new javax.swing.JLabel();
-        jDentistaDialog = new javax.swing.JDialog();
-        jLabel8 = new javax.swing.JLabel();
-        txtCpfMedico = new javax.swing.JTextField();
-        jLabel9 = new javax.swing.JLabel();
-        txtNomeMedico = new javax.swing.JTextField();
-        jLabel10 = new javax.swing.JLabel();
-        txtRegistroMedico = new javax.swing.JTextField();
-        jLabel11 = new javax.swing.JLabel();
-        JCBEspecialidade = new javax.swing.JComboBox();
-        jBAddDentista = new javax.swing.JButton();
-        jConnDialog = new javax.swing.JDialog();
-        jPanel1 = new javax.swing.JPanel();
-        txtPass = new javax.swing.JTextField();
-        txtUser = new javax.swing.JTextField();
-        jLabel12 = new javax.swing.JLabel();
-        connErro = new javax.swing.JLabel();
-        jLabel14 = new javax.swing.JLabel();
-        btConect = new javax.swing.JButton();
-        jLabel15 = new javax.swing.JLabel();
-        jTabbedPane1 = new javax.swing.JTabbedPane();
-        ClientesPanel = new javax.swing.JPanel();
-        jSeparator2 = new javax.swing.JSeparator();
-        jBtAddNCliente = new javax.swing.JButton();
-        jScrollPane2 = new javax.swing.JScrollPane();
-        tbCliente = new javax.swing.JTable();
-        GerenciaPanel = new javax.swing.JPanel();
-        jSeparator4 = new javax.swing.JSeparator();
-        jButton3 = new javax.swing.JButton();
-        jButton5 = new javax.swing.JButton();
-        jSeparator6 = new javax.swing.JSeparator();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        tbFuncionario = new javax.swing.JTable();
-        jScrollPane3 = new javax.swing.JScrollPane();
-        tbDentista = new javax.swing.JTable();
-        ConsultasPanel = new javax.swing.JPanel();
-        jSeparator1 = new javax.swing.JSeparator();
-        jBtAddConsulta = new javax.swing.JButton();
-        jTextField1 = new javax.swing.JTextField();
-        jButton4 = new javax.swing.JButton();
-        jSeparator5 = new javax.swing.JSeparator();
-        jScrollPane4 = new javax.swing.JScrollPane();
-        tbConsulta = new javax.swing.JTable();
-        jMenuBar1 = new javax.swing.JMenuBar();
-        jMenu1 = new javax.swing.JMenu();
-        jMenuItem1 = new javax.swing.JMenuItem();
-        jMenuItem2 = new javax.swing.JMenuItem();
-        jMenu2 = new javax.swing.JMenu();
-        jMenuItem3 = new javax.swing.JMenuItem();
-        jMenuItem4 = new javax.swing.JMenuItem();
-        btConnBanco = new javax.swing.JMenuItem();
-
-        jConsultaDialog.setMinimumSize(new java.awt.Dimension(400, 300));
-
-        jLabel1.setText("CPF");
-
-        jLabel2.setText("Data");
-
-        jBtMarcarConsulta.setText("Adicionar");
-        jBtMarcarConsulta.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtMarcarConsultaActionPerformed(evt);
-            }
-        });
-
-        cbSetRegMed.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
-
-        jLabel.setText("Médico");
-
-        javax.swing.GroupLayout jConsultaDialogLayout = new javax.swing.GroupLayout(jConsultaDialog.getContentPane());
-        jConsultaDialog.getContentPane().setLayout(jConsultaDialogLayout);
-        jConsultaDialogLayout.setHorizontalGroup(
-            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jConsultaDialogLayout.createSequentialGroup()
-                .addGap(37, 37, 37)
-                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jBtMarcarConsulta)
-                    .addGroup(jConsultaDialogLayout.createSequentialGroup()
-                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(jConsultaDialogLayout.createSequentialGroup()
-                                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 27, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(24, 24, 24))
-                            .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                .addGroup(jConsultaDialogLayout.createSequentialGroup()
-                                    .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                    .addGap(20, 20, 20))
-                                .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jConsultaDialogLayout.createSequentialGroup()
-                                    .addComponent(jLabel)
-                                    .addGap(18, 18, 18))))
-                        .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(cbSetRegMed, 0, 182, Short.MAX_VALUE)
-                            .addComponent(txtRegMedic)
-                            .addComponent(txtCpfCliente))))
-                .addContainerGap(130, Short.MAX_VALUE))
-        );
-        jConsultaDialogLayout.setVerticalGroup(
-            jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jConsultaDialogLayout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel1)
-                    .addComponent(txtCpfCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel2)
-                    .addComponent(txtRegMedic, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jConsultaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(cbSetRegMed, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel))
-                .addGap(27, 27, 27)
-                .addComponent(jBtMarcarConsulta)
-                .addContainerGap(134, Short.MAX_VALUE))
-        );
-
-        jClienteDialog.setMinimumSize(new java.awt.Dimension(400, 300));
-
-        jLabel3.setText("CPF");
-
-        jLabel4.setText("Nome");
-
-        jBtAddCliente.setText("Adicionar");
-        jBtAddCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtAddClienteActionPerformed(evt);
-            }
-        });
-
-        jLabel5.setText("Telefone");
-
-        jLabel7.setText("Resultado: ");
-
-        javax.swing.GroupLayout jClienteDialogLayout = new javax.swing.GroupLayout(jClienteDialog.getContentPane());
-        jClienteDialog.getContentPane().setLayout(jClienteDialogLayout);
-        jClienteDialogLayout.setHorizontalGroup(
-            jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jClienteDialogLayout.createSequentialGroup()
-                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jClienteDialogLayout.createSequentialGroup()
-                        .addGap(30, 30, 30)
-                        .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(jLabel4)
-                            .addComponent(jLabel3)
-                            .addComponent(jLabel5))
-                        .addGap(30, 30, 30)
-                        .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addComponent(txtTelCliente, javax.swing.GroupLayout.DEFAULT_SIZE, 152, Short.MAX_VALUE)
-                            .addComponent(txtNomeCliente)
-                            .addComponent(txtCpfCliente1)))
-                    .addGroup(jClienteDialogLayout.createSequentialGroup()
-                        .addGap(21, 21, 21)
-                        .addComponent(jBtAddCliente))
-                    .addGroup(jClienteDialogLayout.createSequentialGroup()
-                        .addContainerGap()
-                        .addComponent(jLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 342, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(48, Short.MAX_VALUE))
-        );
-        jClienteDialogLayout.setVerticalGroup(
-            jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jClienteDialogLayout.createSequentialGroup()
-                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jClienteDialogLayout.createSequentialGroup()
-                        .addGap(23, 23, 23)
-                        .addComponent(jLabel3))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jClienteDialogLayout.createSequentialGroup()
-                        .addContainerGap()
-                        .addComponent(txtCpfCliente1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel4)
-                    .addComponent(txtNomeCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(jClienteDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel5)
-                    .addComponent(txtTelCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(30, 30, 30)
-                .addComponent(jBtAddCliente)
-                .addGap(18, 18, 18)
-                .addComponent(jLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 31, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(99, Short.MAX_VALUE))
-        );
-
-        jDentistaDialog.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-        jDentistaDialog.setMinimumSize(new java.awt.Dimension(400, 300));
-
-        jLabel8.setText("CPF");
-
-        jLabel9.setText("NOME");
-
-        jLabel10.setText("REG");
-
-        jLabel11.setText("ESP");
-
-        jBAddDentista.setText("Adicionar");
-        jBAddDentista.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBAddDentistaActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jDentistaDialogLayout = new javax.swing.GroupLayout(jDentistaDialog.getContentPane());
-        jDentistaDialog.getContentPane().setLayout(jDentistaDialogLayout);
-        jDentistaDialogLayout.setHorizontalGroup(
-            jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jDentistaDialogLayout.createSequentialGroup()
-                .addGap(21, 21, 21)
-                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jDentistaDialogLayout.createSequentialGroup()
-                        .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(jLabel11)
-                            .addComponent(jLabel10)
-                            .addComponent(jLabel9)
-                            .addComponent(jLabel8))
-                        .addGap(28, 28, 28)
-                        .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(txtNomeMedico, javax.swing.GroupLayout.PREFERRED_SIZE, 223, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(txtCpfMedico, javax.swing.GroupLayout.PREFERRED_SIZE, 134, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(txtRegistroMedico, javax.swing.GroupLayout.PREFERRED_SIZE, 87, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(JCBEspecialidade, javax.swing.GroupLayout.PREFERRED_SIZE, 160, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                    .addGroup(jDentistaDialogLayout.createSequentialGroup()
-                        .addGap(11, 11, 11)
-                        .addComponent(jBAddDentista)))
-                .addContainerGap(99, Short.MAX_VALUE))
-        );
-        jDentistaDialogLayout.setVerticalGroup(
-            jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jDentistaDialogLayout.createSequentialGroup()
-                .addGap(39, 39, 39)
-                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel8)
-                    .addComponent(txtCpfMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(txtNomeMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel9))
-                .addGap(18, 18, 18)
-                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(txtRegistroMedico, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel10))
-                .addGap(18, 18, 18)
-                .addGroup(jDentistaDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel11)
-                    .addComponent(JCBEspecialidade, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(41, 41, 41)
-                .addComponent(jBAddDentista)
-                .addContainerGap(63, Short.MAX_VALUE))
-        );
-
-        jConnDialog.setTitle("Banco de Dados");
-        jConnDialog.setMinimumSize(new java.awt.Dimension(400, 300));
-        jConnDialog.setResizable(false);
-
-        jLabel12.setText("Senha");
-
-        connErro.setForeground(new java.awt.Color(255, 0, 0));
-
-        jLabel14.setText("Usuario");
-
-        btConect.setText("Continuar");
-        btConect.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btConectActionPerformed(evt);
-            }
-        });
-
-        jLabel15.setText("Assistente de Conexão com o Banco de Dados");
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(39, 39, 39)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(connErro, javax.swing.GroupLayout.PREFERRED_SIZE, 262, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel14, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel12))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(txtUser, javax.swing.GroupLayout.DEFAULT_SIZE, 114, Short.MAX_VALUE)
-                            .addComponent(txtPass)))
-                    .addComponent(btConect))
-                .addContainerGap(99, Short.MAX_VALUE))
-            .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                    .addContainerGap(73, Short.MAX_VALUE)
-                    .addComponent(jLabel15, javax.swing.GroupLayout.PREFERRED_SIZE, 262, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGap(65, 65, 65)))
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(63, 63, 63)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(txtUser, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel14))
-                .addGap(18, 18, 18)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel12)
-                    .addComponent(txtPass, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(37, 37, 37)
-                .addComponent(btConect)
-                .addGap(29, 29, 29)
-                .addComponent(connErro, javax.swing.GroupLayout.PREFERRED_SIZE, 23, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(67, Short.MAX_VALUE))
-            .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(jPanel1Layout.createSequentialGroup()
-                    .addGap(21, 21, 21)
-                    .addComponent(jLabel15, javax.swing.GroupLayout.PREFERRED_SIZE, 34, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addContainerGap(245, Short.MAX_VALUE)))
-        );
-
-        javax.swing.GroupLayout jConnDialogLayout = new javax.swing.GroupLayout(jConnDialog.getContentPane());
-        jConnDialog.getContentPane().setLayout(jConnDialogLayout);
-        jConnDialogLayout.setHorizontalGroup(
-            jConnDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 400, Short.MAX_VALUE)
-            .addGroup(jConnDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(jConnDialogLayout.createSequentialGroup()
-                    .addGap(0, 0, Short.MAX_VALUE)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGap(0, 0, Short.MAX_VALUE)))
-        );
-        jConnDialogLayout.setVerticalGroup(
-            jConnDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 300, Short.MAX_VALUE)
-            .addGroup(jConnDialogLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(jConnDialogLayout.createSequentialGroup()
-                    .addGap(0, 0, Short.MAX_VALUE)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGap(0, 0, Short.MAX_VALUE)))
-        );
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setBackground(new java.awt.Color(255, 255, 255));
-        setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-        setResizable(false);
-
-        jTabbedPane1.setBackground(java.awt.SystemColor.window);
-
-        ClientesPanel.setBackground(new java.awt.Color(255, 255, 255));
-
-        jBtAddNCliente.setText("Adicionar Cliente");
-        jBtAddNCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtAddNClienteActionPerformed(evt);
-            }
-        });
-
-        tbCliente.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane2.setViewportView(tbCliente);
-
-        javax.swing.GroupLayout ClientesPanelLayout = new javax.swing.GroupLayout(ClientesPanel);
-        ClientesPanel.setLayout(ClientesPanelLayout);
-        ClientesPanelLayout.setHorizontalGroup(
-            ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(ClientesPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(ClientesPanelLayout.createSequentialGroup()
-                        .addGap(10, 10, 10)
-                        .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addGroup(ClientesPanelLayout.createSequentialGroup()
-                        .addGroup(ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jSeparator2)
-                            .addGroup(ClientesPanelLayout.createSequentialGroup()
-                                .addComponent(jBtAddNCliente)
-                                .addGap(0, 514, Short.MAX_VALUE)))
-                        .addContainerGap())))
-        );
-        ClientesPanelLayout.setVerticalGroup(
-            ClientesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(ClientesPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jBtAddNCliente)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jSeparator2, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jScrollPane2)
-                .addContainerGap())
-        );
-
-        jTabbedPane1.addTab("Clientes", ClientesPanel);
-
-        GerenciaPanel.setBackground(new java.awt.Color(255, 255, 255));
-
-        jButton3.setText("Adicionar Dentista");
-        jButton3.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton3ActionPerformed(evt);
-            }
-        });
-
-        jButton5.setText("Adicionar Funcionario");
-
-        tbFuncionario.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane1.setViewportView(tbFuncionario);
-
-        tbDentista.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane3.setViewportView(tbDentista);
-
-        javax.swing.GroupLayout GerenciaPanelLayout = new javax.swing.GroupLayout(GerenciaPanel);
-        GerenciaPanel.setLayout(GerenciaPanelLayout);
-        GerenciaPanelLayout.setHorizontalGroup(
-            GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                        .addGap(10, 10, 10)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addContainerGap(175, Short.MAX_VALUE))
-                    .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                        .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jSeparator4)
-                            .addComponent(jSeparator6, javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                                .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(jButton3)
-                                    .addComponent(jButton5))
-                                .addGap(0, 0, Short.MAX_VALUE)))
-                        .addContainerGap())))
-            .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                    .addGap(20, 20, 20)
-                    .addComponent(jScrollPane3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addContainerGap(175, Short.MAX_VALUE)))
-        );
-        GerenciaPanelLayout.setVerticalGroup(
-            GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jButton3)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jSeparator4, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(192, 192, 192)
-                .addComponent(jButton5)
-                .addGap(5, 5, 5)
-                .addComponent(jSeparator6, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 168, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(60, Short.MAX_VALUE))
-            .addGroup(GerenciaPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(GerenciaPanelLayout.createSequentialGroup()
-                    .addGap(66, 66, 66)
-                    .addComponent(jScrollPane3, javax.swing.GroupLayout.PREFERRED_SIZE, 168, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addContainerGap(285, Short.MAX_VALUE)))
-        );
-
-        jTabbedPane1.addTab("Gerenciar", GerenciaPanel);
-
-        ConsultasPanel.setBackground(new java.awt.Color(255, 255, 255));
-
-        jBtAddConsulta.setText("Adicionar Consulta");
-        jBtAddConsulta.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtAddConsultaActionPerformed(evt);
-            }
-        });
-
-        jButton4.setText("Pesquisar");
-
-        jSeparator5.setOrientation(javax.swing.SwingConstants.VERTICAL);
-
-        tbConsulta.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane4.setViewportView(tbConsulta);
-
-        javax.swing.GroupLayout ConsultasPanelLayout = new javax.swing.GroupLayout(ConsultasPanel);
-        ConsultasPanel.setLayout(ConsultasPanelLayout);
-        ConsultasPanelLayout.setHorizontalGroup(
-            ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(ConsultasPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jSeparator1)
-                    .addGroup(ConsultasPanelLayout.createSequentialGroup()
-                        .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, 586, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(ConsultasPanelLayout.createSequentialGroup()
-                                .addComponent(jBtAddConsulta)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(jSeparator5, javax.swing.GroupLayout.PREFERRED_SIZE, 8, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 243, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(jButton4)))
-                        .addGap(0, 41, Short.MAX_VALUE)))
-                .addContainerGap())
-        );
-        ConsultasPanelLayout.setVerticalGroup(
-            ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(ConsultasPanelLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                        .addGroup(ConsultasPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jButton4))
-                        .addComponent(jSeparator5))
-                    .addComponent(jBtAddConsulta))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, 402, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(61, Short.MAX_VALUE))
-        );
-
-        jTabbedPane1.addTab("Consultas", ConsultasPanel);
-
-        jMenu1.setText("File");
-
-        jMenuItem1.setText("jMenuItem1");
-        jMenu1.add(jMenuItem1);
-
-        jMenuItem2.setText("jMenuItem2");
-        jMenu1.add(jMenuItem2);
-
-        jMenuBar1.add(jMenu1);
-
-        jMenu2.setText("Opções");
-
-        jMenuItem3.setText("Sair");
-        jMenu2.add(jMenuItem3);
-
-        jMenuItem4.setText("Fazer login como cliente");
-        jMenu2.add(jMenuItem4);
-
-        btConnBanco.setText("Banco de Dados");
-        btConnBanco.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btConnBancoActionPerformed(evt);
-            }
-        });
-        jMenu2.add(btConnBanco);
-
-        jMenuBar1.add(jMenu2);
-
-        setJMenuBar(jMenuBar1);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jTabbedPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 20, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jTabbedPane1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jBtAddConsultaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddConsultaActionPerformed
-        this.jConsultaDialog.setVisible(true);
-
-    }//GEN-LAST:event_jBtAddConsultaActionPerformed
-
-    private void jBtMarcarConsultaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtMarcarConsultaActionPerformed
-        Date d = new Date();
-        int reg = cbSetRegMed.getSelectedIndex();
-        if(_clinica.MarcarConsulta(new Consulta(txtCpfCliente.getText(), _clinica.Medicos.get(reg).getRegistro_M(), d)))
-        {
-            this.jConsultaDialog.setVisible(false);
-        }
-
-    }//GEN-LAST:event_jBtMarcarConsultaActionPerformed
-
-    private void jBtAddClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddClienteActionPerformed
-        boolean AddCliente;
-        //AddCliente = _clinica.AddCliente(new Cliente(txtCpfCliente1.getText(), txtNomeCliente.getText(), txtTelCliente.getText()));
-        //jLabel7.setText("" + AddCliente);
-        if (_clinica.AddCliente(new Cliente(txtCpfCliente1.getText(), txtNomeCliente.getText(), txtTelCliente.getText()))) {
-            this.jClienteDialog.setVisible(false);
-        } else {
-            //this.jLabel7.setText(" " + AddCliente);
-        };
-        attTableCliente();
-    }//GEN-LAST:event_jBtAddClienteActionPerformed
-
-    private void jBtAddNClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtAddNClienteActionPerformed
-        this.jClienteDialog.setVisible(true);
-    }//GEN-LAST:event_jBtAddNClienteActionPerformed
-
-    private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton3ActionPerformed
-        this.jDentistaDialog.setVisible(true);
-    }//GEN-LAST:event_jButton3ActionPerformed
-
-    private void jBAddDentistaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBAddDentistaActionPerformed
-        int esp = JCBEspecialidade.getSelectedIndex();
-        esp++;
-        if (_clinica.AddMedico(new Medico(txtCpfMedico.getText(), txtNomeMedico.getText(),
-                txtRegistroMedico.getText(), esp))) {
-            this.jDentistaDialog.setVisible(false);
-        }
-        attTableDentista();
-    }//GEN-LAST:event_jBAddDentistaActionPerformed
-
-    private void btConectActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btConectActionPerformed
-        Connect.setPass(this.txtUser.getText(), this.txtPass.getText());
-        Connection conn = Connect.getConnection();
-        if (conn == null) {
-            connErro.setText("Erro na conexão com o banco de dados!");
-        } else {
-            this.jConnDialog.setVisible(false);
-            attTableCliente();
-        }
-    }//GEN-LAST:event_btConectActionPerformed
-
-    private void btConnBancoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btConnBancoActionPerformed
-        this.jConnDialog.setVisible(true);
-    }//GEN-LAST:event_btConnBancoActionPerformed
-    private void attTableCliente() {
-        //Atualiza a tabela Cliente;
-        DefaultTableModel model = new DefaultTableModel();
-        model.setColumnIdentifiers(Cliente.colunasCliente);
-        _clinica.Clientes = _clinica.MostrarClientes();
-        for (Cliente c : _clinica.Clientes) {
-            model.addRow(new Object[]{c.getCpf(), c.getNome(), c.getTelefone()});
-        }
-        this.tbCliente.setModel(model);
-        this.tbCliente.setFillsViewportHeight(true);
-    }
-
-    private void attTableDentista() {
-        DefaultTableModel model = new DefaultTableModel();
-        model.setColumnIdentifiers(Medico.colunasMedico);
-        _clinica.Medicos = _clinica.MostrarMedicos();
-        for (Medico m : _clinica.Medicos) {
-            model.addRow(new Object[]{m.getCpf_M(), m.getNome_M(), m.getRegistro_M(), m.getStrEspecialidade()});
-        }
-        this.tbDentista.setModel(model);
-        this.tbDentista.setFillsViewportHeight(true);
-        setComboBoxItems();
-    }
-
-    private void attTableConsulta() {
-        DefaultTableModel model = new DefaultTableModel();
-        model.setColumnIdentifiers(Consulta.colunasConsulta);
-        _clinica.Consultas = _clinica.MostrarConsultas();
-        for (Consulta cs : _clinica.Consultas) {
-            model.addRow(new Object[]{cs.getIdConsulta(), cs.getCpfCliente(), cs.getRegistroMedico(), cs.getData()});
-        }
-        this.tbConsulta.setModel(model);
-        this.tbConsulta.setFillsViewportHeight(true);
-    }
-
-    private void setComboBoxItems() {
-
-        //seta o combobox da consulta
-        DefaultComboBoxModel model = new DefaultComboBoxModel();
-        for (Medico m : _clinica.Medicos) {
-            model.addElement(m.getNome_M()+", "+m.getStrEspecialidade());
-        }
-        this.cbSetRegMed.setModel(model);
-        
-        //seta o combobox do addmedico
-        DefaultComboBoxModel dmodel = new DefaultComboBoxModel();
-        for (String m: _clinica.getEspecialidades()){
-            dmodel.addElement(m);
-        }
-        this.JCBEspecialidade.setModel(dmodel);
-    }
-
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Windows".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(ClinicaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(ClinicaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(ClinicaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(ClinicaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        /* Create and display the form */
-
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                //new ClinicaFrame().setVisible(true);
-                ClinicaFrame clinica = new ClinicaFrame();
-                clinica.setVisible(true);
-                clinica.attTableDentista();
-                clinica.attTableCliente();
-                clinica.attTableConsulta();
-                clinica.setComboBoxItems();
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JPanel ClientesPanel;
-    private javax.swing.JPanel ConsultasPanel;
-    private javax.swing.JPanel GerenciaPanel;
-    private javax.swing.JComboBox JCBEspecialidade;
-    private javax.swing.JButton btConect;
-    private javax.swing.JMenuItem btConnBanco;
-    private javax.swing.JComboBox cbSetRegMed;
-    private javax.swing.JLabel connErro;
-    private javax.swing.JButton jBAddDentista;
-    private javax.swing.JButton jBtAddCliente;
-    private javax.swing.JButton jBtAddConsulta;
-    private javax.swing.JButton jBtAddNCliente;
-    private javax.swing.JButton jBtMarcarConsulta;
-    private javax.swing.JButton jButton3;
-    private javax.swing.JButton jButton4;
-    private javax.swing.JButton jButton5;
-    private javax.swing.JDialog jClienteDialog;
-    private javax.swing.JDialog jConnDialog;
-    private javax.swing.JDialog jConsultaDialog;
-    private javax.swing.JDialog jDentistaDialog;
-    private javax.swing.JLabel jLabel;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel10;
-    private javax.swing.JLabel jLabel11;
-    private javax.swing.JLabel jLabel12;
-    private javax.swing.JLabel jLabel14;
-    private javax.swing.JLabel jLabel15;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JLabel jLabel9;
-    private javax.swing.JMenu jMenu1;
-    private javax.swing.JMenu jMenu2;
-    private javax.swing.JMenuBar jMenuBar1;
-    private javax.swing.JMenuItem jMenuItem1;
-    private javax.swing.JMenuItem jMenuItem2;
-    private javax.swing.JMenuItem jMenuItem3;
-    private javax.swing.JMenuItem jMenuItem4;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JScrollPane jScrollPane2;
-    private javax.swing.JScrollPane jScrollPane3;
-    private javax.swing.JScrollPane jScrollPane4;
-    private javax.swing.JSeparator jSeparator1;
-    private javax.swing.JSeparator jSeparator2;
-    private javax.swing.JSeparator jSeparator4;
-    private javax.swing.JSeparator jSeparator5;
-    private javax.swing.JSeparator jSeparator6;
-    private javax.swing.JTabbedPane jTabbedPane1;
-    private javax.swing.JTextField jTextField1;
-    private javax.swing.JTable tbCliente;
-    private javax.swing.JTable tbConsulta;
-    private javax.swing.JTable tbDentista;
-    private javax.swing.JTable tbFuncionario;
-    private javax.swing.JTextField txtCpfCliente;
-    private javax.swing.JTextField txtCpfCliente1;
-    private javax.swing.JTextField txtCpfMedico;
-    private javax.swing.JTextField txtNomeCliente;
-    private javax.swing.JTextField txtNomeMedico;
-    private javax.swing.JTextField txtPass;
-    private javax.swing.JTextField txtRegMedic;
-    private javax.swing.JTextField txtRegistroMedico;
-    private javax.swing.JTextField txtTelCliente;
-    private javax.swing.JTextField txtUser;
-    // End of variables declaration//GEN-END:variables
-}
