+++ Clinica_Odonto/src/Users/Cliente.java
@@ -1,5 +1,7 @@
 package Users;
 
+import java.util.ArrayList;
+
 /**
  *
  * @author LucasPc
@@ -9,11 +11,26 @@ public class Cliente {
     private String cpf;
     private String nome;
     private String telefone;
-    public static String colunasCliente[] = {"CPF", "NOME", "TEL"};
+    public ArrayList<Consulta> consultas = new ArrayList<>();
+    public static String colunasCliente[] = {"Cpf", "Nome", "Tel"};
+    public static String colConsultasCliente[] = {"ID", "Data", "Medico"};
     public Cliente() {
 
     }
 
+    public static String[] getColunasCliente() {
+        return colunasCliente;
+    }
+
+    public static void setColunasCliente(String[] colunasCliente) {
+        Cliente.colunasCliente = colunasCliente;
+    }
+
+    public ArrayList<Consulta> getConsultas() {
+        return new Clinica().PesquisarConsulta(this.getCpf());
+    }
+
+    
     public Cliente(String Cpf, String Nome, String Tel) {
         this.cpf = Cpf;
         this.nome = Nome;
