+++ Clinica_Odonto/src/Users/Consulta.java
@@ -1,4 +1,3 @@
-
 package Users;
 
 import java.util.*;
@@ -8,38 +7,31 @@ import java.util.*;
  * @author LucasPc
  */
 public class Consulta {
-    
-    public static String colunasConsulta[] = {"ID", "CLIENTE", "MEDICO", "DATA"};
+
+    public static String colunasConsulta[] = {"Id", "Cliente", "Medico", "Data"};
     private String cpfCliente;
-    private String registroMedico;
+    public Medico dadosMedico;
     private int idConsulta;
     private Date data = new Date();
 
     public Consulta(String _cpf, String _registro, Date _data) {
+        dadosMedico = new Medico();
         this.cpfCliente = _cpf;
-        this.registroMedico = _registro;
+        this.dadosMedico.setRegistro_M(_registro);
         this.data = _data;
         idConsulta = new Random().nextInt(1000);
     }
 
     public Consulta() {
+        dadosMedico = new Medico();
     }
 
-    
     public String getCpfCliente() {
         return this.cpfCliente;
     }
 
     public void setCpfCliente(String _cpf) {
         this.cpfCliente = _cpf;
-    }
-
-    public String getRegistroMedico() {
-        return this.registroMedico;
-    }
-
-    public void setRegistroMedico(String _registro) {
-        this.registroMedico = _registro;
     }
 
     public int getIdConsulta() {
